refactored MapSearch.js to use redux
